fix: channels for cep and float min flight time

+++ model/src/main/java/rs/moma/flightforge/model/UserPreferences.java
@@ -11,7 +11,7 @@ public class UserPreferences {
     private double foamboardWeight;
     private double scaleFactor;
     private Double minTWRatio;
-    private Integer minFlightTime;
+    private Double minFlightTime;
     private Priority priority;
     private boolean metalGearsPreference;
     private String location;
